// src/lib/enums/product.enum.ts
export enum ProductVolume {
  THIRTY = 30,
  FIFTY = 50,
  HUNDRED = 100,
}

export enum ProductStatus {
  ACTIVE = "ACTIVE",
  INACTIVE = "INACTIVE",
  OUTOFSTOCK = "OUTOFSTOCK",
}

export enum ProductGender {
  MEN = "MEN",
  WOMEN = "WOMEN",
  UNISEX = "UNISEX",
}

export enum ProductType {
  EAU_DE_PARFUM = "EAU_DE_PARFUM",
  EAU_DE_TOILETTE = "EAU_DE_TOILETTE",
  PARFUM = "PARFUM",
  BODY_SPRAY = "BODY_SPRAY",
}

export enum ProductFragrance {
  FLORAL = "FLORAL",
  WOODY = "WOODY",
  CITRUS = "CITRUS",
  ORIENTAL = "ORIENTAL",
  AROMATIC = "AROMATIC",
  FRESH = "FRESH",
}